refactor(core): split domain/agent.rs into agent/ submodule directory
Move AgentConfig, tool_types, messages, and events into separate files.
Also moves prune_keep_* constants from hardcoded values in context_pruning
into AgentConfig so they are configurable and visible alongside other loop
parameters.

No callers change — domain/agent/mod.rs re-exports everything.
