+++ crates/gglib-core/src/domain/agent/events.rs
@@ -1,0 +1,139 @@
+//! [`AgentEvent`] and [`LlmStreamEvent`] — observable events in the agentic loop.
+
+use serde::{Deserialize, Serialize};
+
+use super::tool_types::{ToolCall, ToolResult};
+
+// =============================================================================
+// Agent events (SSE stream units)
+// =============================================================================
+
+/// An observable event emitted by the agentic loop.
+///
+/// These events are the unit of SSE emission: every state change in the loop
+/// produces exactly one variant. Axum SSE handlers serialise these to
+/// `data: <json>\n\n` frames; CLI consumers may log or render them directly.
+///
+/// # Serde tag
+///
+/// `#[serde(tag = "type", rename_all = "snake_case")]` produces e.g.
+/// `{"type":"tool_call_start","tool_call":{...}}`.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(tag = "type", rename_all = "snake_case")]
+pub enum AgentEvent {
+    /// An incremental text fragment from the model's response.
+    TextDelta {
+        /// The new text fragment (append to the current buffer).
+        content: String,
+    },
+
+    /// The model has requested execution of a tool.
+    ToolCallStart {
+        /// The tool call that is about to be dispatched.
+        tool_call: ToolCall,
+    },
+
+    /// A tool execution has completed (success or failure).
+    ToolCallComplete {
+        /// The outcome of the tool, including timing and success flag.
+        result: ToolResult,
+    },
+
+    /// One full LLM→tool-execution cycle has completed.
+    IterationComplete {
+        /// The 1-based iteration index that just finished.
+        iteration: usize,
+        /// Number of tool calls executed during this iteration.
+        tool_calls: usize,
+    },
+
+    /// The loop has concluded and produced a definitive answer.
+    FinalAnswer {
+        /// The complete final response text.
+        content: String,
+    },
+
+    /// A fatal error has terminated the loop.
+    Error {
+        /// Human-readable description of the failure.
+        message: String,
+    },
+}
+
+// =============================================================================
+// LLM stream events (consumed by LlmCompletionPort implementors)
+// =============================================================================
+
+/// A single event produced by a streaming LLM response.
+///
+/// These low-level events are the currency of [`crate::ports::LlmCompletionPort`];
+/// they are parsed by adapter crates from raw SSE frames and handed to
+/// `gglib-agent`'s stream collector, which:
+///
+/// - Forwards [`TextDelta`](LlmStreamEvent::TextDelta) items directly to the
+///   caller's [`AgentEvent`] channel so text appears in real time.
+/// - Accumulates [`ToolCallDelta`](LlmStreamEvent::ToolCallDelta) fragments
+///   until the stream ends, then assembles them into [`ToolCall`] values.
+/// - Waits for [`Done`](LlmStreamEvent::Done) before triggering tool execution.
+#[derive(Debug, Clone)]
+pub enum LlmStreamEvent {
+    /// An incremental text fragment from the model's response.
+    TextDelta {
+        /// The new text fragment (append to the running content buffer).
+        content: String,
+    },
+
+    /// An incremental fragment of a tool-call request.
+    ///
+    /// The adapter crate streams these before the model has finished
+    /// generating the full arguments JSON. The stream collector accumulates
+    /// all deltas for a given `index` into a single [`ToolCall`].
+    ToolCallDelta {
+        /// Zero-based index of the tool call within the current response.
+        index: usize,
+        /// Call identifier (only present in the first delta for this index).
+        id: Option<String>,
+        /// Tool name (only present in the first delta for this index).
+        name: Option<String>,
+        /// Partial arguments JSON string fragment (accumulate with `push_str`).
+        arguments: Option<String>,
+    },
+
+    /// Signals the end of the stream.
+    ///
+    /// Every conforming stream must end with exactly one `Done` item.
+    Done {
+        /// The OpenAI-compatible finish reason (e.g. `"stop"`, `"tool_calls"`,
+        /// `"length"`).
+        finish_reason: String,
+    },
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn agent_event_serde_tag_matches_wire_format() {
+        let evt = AgentEvent::FinalAnswer {
+            content: "done".into(),
+        };
+        let json = serde_json::to_value(&evt).unwrap();
+        assert_eq!(json["type"], "final_answer");
+        assert_eq!(json["content"], "done");
+    }
+
+    #[test]
+    fn tool_call_start_serialises_correctly() {
+        let evt = AgentEvent::ToolCallStart {
+            tool_call: ToolCall {
+                id: "c1".into(),
+                name: "search".into(),
+                arguments: serde_json::json!({"q": "rust"}),
+            },
+        };
+        let json = serde_json::to_value(&evt).unwrap();
+        assert_eq!(json["type"], "tool_call_start");
+        assert_eq!(json["tool_call"]["name"], "search");
+    }
+}
