+++ crates/gglib-core/src/domain/agent/messages.rs
@@ -1,0 +1,86 @@
+//! [`AgentMessage`] — A single message in the agent conversation.
+
+use serde::{Deserialize, Serialize};
+
+use super::tool_types::ToolCall;
+
+/// A single message in the agent conversation.
+///
+/// The closed enum prevents invalid states that a flat struct with `role: String`
+/// would allow (e.g. a `User` message carrying `tool_calls`, or a `Tool` message
+/// without a `tool_call_id`).
+///
+/// # Wire format
+///
+/// `#[serde(tag = "role", rename_all = "lowercase")]` produces JSON identical to
+/// the TypeScript `ChatMessage` interface in the frontend:
+///
+/// ```json
+/// { "role": "user", "content": "What files are in the project?" }
+/// { "role": "assistant", "content": null, "tool_calls": [...] }
+/// { "role": "tool", "tool_call_id": "call_abc", "content": "src/\nlib/" }
+/// ```
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(tag = "role", rename_all = "lowercase")]
+pub enum AgentMessage {
+    /// A system-level instruction that sets the model's persona and constraints.
+    System {
+        /// Instruction text.
+        content: String,
+    },
+
+    /// A message from the human user.
+    User {
+        /// Message text.
+        content: String,
+    },
+
+    /// A response from the assistant model.
+    ///
+    /// Either `content` **or** `tool_calls` is non-`None`; both may be present
+    /// when the model produces a reasoning preamble before requesting tool calls.
+    Assistant {
+        /// Optional text content of the response.
+        #[serde(skip_serializing_if = "Option::is_none")]
+        content: Option<String>,
+
+        /// Tool calls requested by the model (triggers the tool execution phase).
+        #[serde(skip_serializing_if = "Option::is_none")]
+        tool_calls: Option<Vec<ToolCall>>,
+    },
+
+    /// The result of a tool call, to be sent back to the model.
+    Tool {
+        /// Must match the [`ToolCall::id`] from the preceding `Assistant` message.
+        tool_call_id: String,
+
+        /// Serialised output of the tool (or error description if it failed).
+        content: String,
+    },
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn serde_tag_matches_wire_format() {
+        let msg = AgentMessage::Tool {
+            tool_call_id: "call_1".into(),
+            content: "ok".into(),
+        };
+        let json = serde_json::to_value(&msg).unwrap();
+        assert_eq!(json["role"], "tool");
+        assert_eq!(json["tool_call_id"], "call_1");
+    }
+
+    #[test]
+    fn assistant_with_no_tool_calls_omits_field() {
+        let msg = AgentMessage::Assistant {
+            content: Some("hi".into()),
+            tool_calls: None,
+        };
+        let json = serde_json::to_value(&msg).unwrap();
+        assert!(json.get("tool_calls").is_none());
+    }
+}
