+++ crates/gglib-core/src/domain/agent/tool_types.rs
@@ -1,0 +1,140 @@
+//! Tool-schema and tool-call/result types.
+
+use serde::{Deserialize, Serialize};
+
+// =============================================================================
+// Tool schema
+// =============================================================================
+
+/// A tool that the LLM may invoke, expressed as a name + JSON Schema.
+///
+/// This is the agent domain's own type. Adapter layers are responsible for
+/// converting their internal tool representations:
+///
+/// - MCP adapter: `McpTool → ToolDefinition`
+/// - Built-in tools: constructed directly
+///
+/// The agent domain must **not** reference `McpTool` or any crate-external
+/// tool type directly.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ToolDefinition {
+    /// Unique function name exposed to the LLM (e.g. `"filesystem_read_file"`).
+    pub name: String,
+
+    /// Human-readable description sent to the LLM to explain when to use the tool.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+
+    /// JSON Schema object describing the tool's input parameters.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub input_schema: Option<serde_json::Value>,
+}
+
+impl ToolDefinition {
+    /// Create a minimal tool definition with only a name.
+    #[must_use]
+    pub fn new(name: impl Into<String>) -> Self {
+        Self {
+            name: name.into(),
+            description: None,
+            input_schema: None,
+        }
+    }
+
+    /// Attach a human-readable description.
+    #[must_use]
+    pub fn with_description(mut self, description: impl Into<String>) -> Self {
+        self.description = Some(description.into());
+        self
+    }
+
+    /// Attach a JSON Schema for the tool's input parameters.
+    #[must_use]
+    pub fn with_input_schema(mut self, schema: serde_json::Value) -> Self {
+        self.input_schema = Some(schema);
+        self
+    }
+}
+
+// =============================================================================
+// Tool call / result
+// =============================================================================
+
+/// A tool invocation requested by the LLM.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ToolCall {
+    /// Unique identifier for this call, generated by the LLM.
+    ///
+    /// Must be echoed back in the corresponding [`ToolResult::tool_call_id`].
+    pub id: String,
+
+    /// Name of the tool to invoke (matches [`ToolDefinition::name`]).
+    pub name: String,
+
+    /// Arguments as a JSON object, matching the tool's `input_schema`.
+    pub arguments: serde_json::Value,
+}
+
+/// The outcome of executing a [`ToolCall`].
+///
+/// # Important: failures are context, not errors
+///
+/// A `ToolResult` with `success: false` is **not** a Rust-level error —
+/// it is domain data that gets appended to the conversation as an
+/// `AgentMessage::Tool` entry so the model can observe the failure and
+/// adjust its strategy (retry, fix arguments, use a different tool, etc.).
+///
+/// Infrastructure-level failures (e.g. the MCP process crashed) are
+/// surfaced separately via `anyhow::Error` from [`crate::ports::ToolExecutorPort::execute`]
+/// and then converted by the loop implementation into a `ToolResult` with
+/// `success: false` and an appropriate `content` describing the outage.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ToolResult {
+    /// Echoes the [`ToolCall::id`] this result corresponds to.
+    pub tool_call_id: String,
+
+    /// Human-readable output (or error message) from the tool execution.
+    pub content: String,
+
+    /// Whether the tool completed successfully.
+    ///
+    /// `false` here is **not** a loop error — see type-level docs above.
+    pub success: bool,
+
+    /// Time spent waiting for a concurrency permit (semaphore), in
+    /// milliseconds.
+    pub wait_ms: u64,
+
+    /// Wall-clock time taken to execute the tool (after acquiring the permit),
+    /// in milliseconds.
+    pub duration_ms: u64,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use serde_json::json;
+
+    #[test]
+    fn tool_definition_builder_sets_fields() {
+        let tool = ToolDefinition::new("search")
+            .with_description("Full-text search")
+            .with_input_schema(json!({ "type": "object" }));
+        assert_eq!(tool.name, "search");
+        assert!(tool.description.is_some());
+        assert!(tool.input_schema.is_some());
+    }
+
+    #[test]
+    fn tool_result_success_false_is_serialisable() {
+        let result = ToolResult {
+            tool_call_id: "c1".into(),
+            content: "ERROR: file not found".into(),
+            success: false,
+            wait_ms: 0,
+            duration_ms: 12,
+        };
+        let json = serde_json::to_value(&result).unwrap();
+        assert_eq!(json["success"], false);
+    }
+}
